requirements.txt
absl-py==0.13.0
apex==0.1
astunparse==1.6.3
batchgenerators==0.23
blessings==1.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.2
clang==5.0
cloudpickle==2.0.0
cycler==0.10.0
decorator==5.1.0
dm-tree==0.1.6
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.1
h5py==3.1.0
idna==3.2
imageio==2.9.0
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
linecache2==1.0.0
Markdown==3.3.4
matplotlib==3.4.3
networkx==1.8.1
nibabel==3.2.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.2.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2021.8.30
torch==1.7.1+cu101
torch-ema @ git+https://github.com/fadel/pytorch_ema@3985995e523aa25dd3cff7e7984130eef90a4282
torch-summary==1.4.5
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.62.3
traceback2==1.4.0
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
